requirements.txt
fastapi
sqlalchemy
gtts
pillow
chardet
langdetect
uvicorn
